requirements.txt
appdirs==1.4.4
attrs==23.1.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
diff-cover==7.6.0
exceptiongroup==1.1.1
fastf1==3.0.4
fonttools==4.40.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.0.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.3.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-cache==1.0.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlfluff==2.1.1
tblib==1.7.0
thefuzz==0.19.0
timple==0.1.6
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
url-normalize==1.4.3
urllib3==1.26.16
websockets==11.0.3
zipp==3.15.0
